moved event to features folder, added no event message and fixed contact
