+++ src/app/(frontend)/contact/_components/ContactForm.tsx
@@ -10,9 +10,6 @@ import FormTextarea from '@/components/ui/FormTextArea';
 import { Button } from '@/components/ui/Button';
 import ThankYou from './ThankYou';
 
-// T3 env
-import { env } from 'config/clientEnv';
-
 export default function ContactForm() {
     const [sent, setSent] = useState(false);
 
@@ -33,7 +30,7 @@ export default function ContactForm() {
                     'Content-Type': 'application/json',
                 },
                 body: JSON.stringify({
-                    access_key: env.NEXT_PUBLIC_WEB3FORMS_KEY,
+                    access_key: "869b1607-6451-4661-b68c-27a1738c0bcc",
                     name: values.name,
                     email: values.email,
                     message: values.message,
