+++ src/features/events/components/EventInfo.tsx
