+++ src/features/events/components/NoEvents.tsx
@@ -1,0 +1,8 @@
+export default function NoEvents() {
+    return (
+        <div className="text-center">
+            <h3>No Upcoming Events</h3>
+            <p>Please check back later, or follow our Instagram to be notified on our next event!</p>
+        </div>
+    );
+}
